12629: 引数から設定値を渡せるよう修正
