+++ src/server.ts
@@ -27,7 +27,7 @@ export class ProxyServer extends EventEmitter {
         super();
     }
 
-    public async run(serverPort: number): Promise<number> {
+    public async run(serverPort: number, proxySettings: IIOSProxySettings, deviceTarget: IIOSDeviceTarget): Promise<number> {
         this._serverPort = serverPort;
         this._clients = new Map<ws, string>();
 
@@ -44,31 +44,15 @@ export class ProxyServer extends EventEmitter {
 
         // Start server and return the port number
         this._hs.listen(this._serverPort);
-        const port = this._hs.address().port;
+        const port = this._serverPort;
 
         // const settings = await IOSAdapter.getProxySettings({
         //     proxyPath: null,
         //     proxyPort: (port + 100),
         //     proxyArgs: null
         // });
-        const proxyPort = port + 100
-        const deviceId = '3fd3ef5648eeed4795e22ed002f6e6fcd2d32b3b';
-        const settings = {
-            proxyPath: '/usr/local/bin/ios_webkit_debug_proxy',
-            proxyPort: proxyPort,
-            proxyArgs: [
-                '--no-frontend',
-                '--config=' + deviceId + ':' + proxyPort
-            ]
-        };
-        const deviceTarget = {
-            deviceId: deviceId,
-            deviceName: deviceId,
-            url: 'localhost:' + proxyPort,
-            version: ''
-        };
 
-        this._adapter = new IOSAdapter(`/ios`, `ws://localhost:${port}`, <IIOSProxySettings>settings, <IIOSDeviceTarget> deviceTarget);
+        this._adapter = new IOSAdapter(`/ios`, `ws://localhost:${port}`, <IIOSProxySettings>proxySettings, <IIOSDeviceTarget> deviceTarget);
         
         return this._adapter.start().then(() => {
             this.startTargetFetcher();
